refactor(api): 移动 Objective 列表响应模型
Why:
- Objective 分页响应定义在路由模块中，而 schema 模块仍保留未使用的非分页列表模型。

What:
- 将排名与分页响应模型迁入 Core schema 模块，删除旧列表模型，并让路由直接导入分页契约。

Impact:
- Objective 列表的 JSON 字段、分页行为和 OpenAPI 组件名称保持不变。

Verification:
- Objective 路由的 9 项测试全部通过。
- OpenAPI 仍生成原有的排名与分页响应组件。
